add smj support
