+++ gluten-core/src/main/scala/io/glutenproject/execution/JoinUtils.scala
@@ -1,0 +1,277 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.glutenproject.execution
+
+import com.google.protobuf.Any
+import io.glutenproject.expression.{AttributeReferenceTransformer, ConverterUtils, ExpressionConverter}
+import io.glutenproject.substrait.expression.{ExpressionBuilder, ExpressionNode}
+import io.glutenproject.substrait.rel.{RelBuilder, RelNode}
+import io.glutenproject.substrait.SubstraitContext
+import io.glutenproject.substrait.`type`.{TypeBuilder, TypeNode}
+import io.glutenproject.substrait.extensions.{AdvancedExtensionNode, ExtensionBuilder}
+import io.substrait.proto.JoinRel
+import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeReference, Expression}
+import org.apache.spark.sql.catalyst.plans.{ExistenceJoin, FullOuter, InnerLike, JoinType, LeftExistence, LeftOuter, RightOuter}
+import org.apache.spark.sql.types.DataType
+
+import java.util
+import scala.collection.JavaConverters._
+
+object JoinUtils {
+  private def createEnhancement(output: Seq[Attribute]): com.google.protobuf.Any = {
+    val inputTypeNodes = output.map { attr =>
+      ConverterUtils.getTypeNode(attr.dataType, attr.nullable)
+    }
+    // Normally the enhancement node is only used for plan validation. But here the enhancement
+    // is also used in execution phase. In this case an empty typeUrlPrefix need to be passed,
+    // so that it can be correctly parsed into json string on the cpp side.
+    Any.pack(TypeBuilder.makeStruct(false,
+      new util.ArrayList[TypeNode](inputTypeNodes.asJava)).toProtobuf,
+      /* typeUrlPrefix */"")
+  }
+
+  def createExtensionNode(output: Seq[Attribute],
+                          validation: Boolean): AdvancedExtensionNode = {
+    // Use field [enhancement] in a extension node for input type validation.
+    if (validation) {
+      ExtensionBuilder.makeAdvancedExtension(createEnhancement(output))
+    } else {
+      null
+    }
+  }
+
+  def preProjectionNeeded(keyExprs: Seq[Expression]): Boolean = {
+    !keyExprs.forall(_.isInstanceOf[AttributeReference])
+  }
+
+  def createPreProjectionIfNeeded(keyExprs: Seq[Expression],
+                                  inputNode: RelNode,
+                                  inputNodeOutput: Seq[Attribute],
+                                  joinOutput: Seq[Attribute],
+                                  substraitContext: SubstraitContext,
+                                  operatorId: java.lang.Long,
+                                  validation: Boolean)
+  : (Seq[(ExpressionNode, DataType)], RelNode, Seq[Attribute]) = {
+    if (!preProjectionNeeded(keyExprs)) {
+      // Skip pre-projection if all keys are [AttributeReference]s,
+      // which can be directly converted into SelectionNode.
+      val keys = keyExprs.map { expr =>
+        (ExpressionConverter.replaceWithExpressionTransformer(expr, joinOutput)
+          .asInstanceOf[AttributeReferenceTransformer]
+          .doTransform(substraitContext.registeredFunction), expr.dataType)
+      }
+      (keys, inputNode, inputNodeOutput)
+    } else {
+      // Pre-projection is constructed from original columns followed by join-key expressions.
+      val selectOrigins = inputNodeOutput.indices.map(ExpressionBuilder.makeSelection(_))
+      val appendedKeys = keyExprs.flatMap {
+        case _: AttributeReference => None
+        case expr =>
+          Some(
+            (ExpressionConverter
+              .replaceWithExpressionTransformer(expr, inputNodeOutput)
+              .doTransform(substraitContext.registeredFunction), expr.dataType))
+      }
+      val preProjectNode = RelBuilder.makeProjectRel(
+        inputNode,
+        new java.util.ArrayList[ExpressionNode](
+          (selectOrigins ++ appendedKeys.map(_._1)).asJava),
+        createExtensionNode(inputNodeOutput, validation),
+        substraitContext,
+        operatorId)
+
+      // Compute index for join keys in join outputs.
+      val offset = joinOutput.size - inputNodeOutput.size + selectOrigins.size
+      val appendedKeysAndIndices = appendedKeys.zipWithIndex.iterator
+      val keys = keyExprs.map {
+        case a: AttributeReference =>
+          // The selection index for original AttributeReference is unchanged.
+          (ExpressionConverter.replaceWithExpressionTransformer(a, joinOutput)
+            .asInstanceOf[AttributeReferenceTransformer]
+            .doTransform(substraitContext.registeredFunction), a.dataType)
+        case _ =>
+          val (key, idx) = appendedKeysAndIndices.next()
+          (ExpressionBuilder.makeSelection(idx + offset), key._2)
+      }
+      (keys, preProjectNode, inputNodeOutput ++
+        appendedKeys.zipWithIndex.map { case (key, idx) =>
+          // Create output attributes for appended keys.
+          // This is used as place holder for finding the right column indexes in post-join filters.
+          AttributeReference(s"col_${idx + offset}", key._2)()
+        }
+      )
+    }
+  }
+
+  def createJoinExtensionNode(joinParameters: Any.Builder,
+                              output: Seq[Attribute]): AdvancedExtensionNode = {
+    // Use field [optimization] in a extension node
+    // to send some join parameters through Substrait plan.
+    val enhancement = createEnhancement(output)
+    ExtensionBuilder.makeAdvancedExtension(joinParameters.build(), enhancement)
+  }
+
+  // Return the direct join output.
+  protected def getDirectJoinOutput(joinType: JoinType,
+                                     leftOutput: Seq[Attribute],
+                                    rightOutput: Seq[Attribute])
+  : (Seq[Attribute], Seq[Attribute]) = {
+    joinType match {
+      case _: InnerLike =>
+        (leftOutput, rightOutput)
+      case LeftOuter =>
+        (leftOutput, rightOutput.map(_.withNullability(true)))
+      case RightOuter =>
+        (leftOutput.map(_.withNullability(true)), rightOutput)
+      case FullOuter =>
+        (leftOutput.map(_.withNullability(true)), rightOutput.map(_.withNullability(true)))
+      case j: ExistenceJoin =>
+        (leftOutput :+ j.exists, Nil)
+      case LeftExistence(_) =>
+        // LeftSemi | LeftAnti | ExistenceJoin.
+        (leftOutput, Nil)
+      case x =>
+        throw new IllegalArgumentException(
+          s"${
+            getClass.getSimpleName
+          } not take $x as the JoinType")
+    }
+  }
+
+  protected def getDirectJoinOutputSeq(joinType: JoinType,
+                                        leftOutput: Seq[Attribute],
+                                       rightOutput: Seq[Attribute]): Seq[Attribute] = {
+    val (left, right) = getDirectJoinOutput(joinType, leftOutput, rightOutput)
+    left ++ right
+  }
+
+  // scalastyle:off argcount
+  def createJoinRel(streamedKeyExprs: Seq[Expression],
+                    buildKeyExprs: Seq[Expression],
+                    condition: Option[Expression],
+                    substraitJoinType: JoinRel.JoinType,
+                    exchangeTable: Boolean,
+                    joinType: JoinType,
+                    joinParameters: Any.Builder,
+                    inputStreamedRelNode: RelNode,
+                    inputBuildRelNode: RelNode,
+                    inputStreamedOutput: Seq[Attribute],
+                    inputBuildOutput: Seq[Attribute],
+                    substraitContext: SubstraitContext,
+                    operatorId: java.lang.Long,
+                    validation: Boolean = false): RelNode = {
+    // scalastyle:on argcount
+    // Create pre-projection for build/streamed plan. Append projected keys to each side.
+    val (streamedKeys, streamedRelNode, streamedOutput) = createPreProjectionIfNeeded(
+      streamedKeyExprs,
+      inputStreamedRelNode,
+      inputStreamedOutput,
+      inputStreamedOutput,
+      substraitContext,
+      operatorId,
+      validation)
+
+    val (buildKeys, buildRelNode, buildOutput) = createPreProjectionIfNeeded(
+      buildKeyExprs,
+      inputBuildRelNode,
+      inputBuildOutput,
+      streamedOutput ++ inputBuildOutput,
+      substraitContext,
+      operatorId,
+      validation)
+
+    // Combine join keys to make a single expression.
+    val joinExpressionNode = (streamedKeys zip buildKeys).map {
+      case ((leftKey, leftType), (rightKey, rightType)) =>
+        HashJoinLikeExecTransformer.makeEqualToExpression(
+          leftKey, leftType, rightKey, rightType, substraitContext.registeredFunction)
+    }.reduce((l, r) =>
+      HashJoinLikeExecTransformer.makeAndExpression(l, r, substraitContext.registeredFunction))
+
+    // Create post-join filter, which will be computed in hash join.
+    val postJoinFilter = condition.map {
+      expr =>
+        ExpressionConverter
+          .replaceWithExpressionTransformer(expr, streamedOutput ++ buildOutput)
+          .doTransform(substraitContext.registeredFunction)
+    }
+
+    // Create JoinRel.
+    val joinRel = RelBuilder.makeJoinRel(
+      streamedRelNode,
+      buildRelNode,
+      substraitJoinType,
+      joinExpressionNode,
+      postJoinFilter.orNull,
+      createJoinExtensionNode(joinParameters, streamedOutput ++ buildOutput),
+      substraitContext,
+      operatorId)
+
+    // Result projection will drop the appended keys, and exchange columns order if BuildLeft.
+    val resultProjection = if (exchangeTable) {
+      val (leftOutput, rightOutput) =
+        getDirectJoinOutput(joinType, inputBuildOutput, inputStreamedOutput)
+      joinType match {
+        case _: ExistenceJoin =>
+          inputBuildOutput.indices.map(ExpressionBuilder.makeSelection(_)) ++
+            Seq(ExpressionBuilder.makeSelection(buildOutput.size))
+        case LeftExistence(_) =>
+          leftOutput.indices.map(ExpressionBuilder.makeSelection(_))
+        case _ =>
+          // Exchange the order of build and streamed.
+          leftOutput.indices.map(idx =>
+            ExpressionBuilder.makeSelection(idx + streamedOutput.size)) ++
+            rightOutput.indices
+              .map(ExpressionBuilder.makeSelection(_))
+      }
+    } else {
+      val (leftOutput, rightOutput) =
+        getDirectJoinOutput(joinType, inputStreamedOutput, inputBuildOutput)
+      if (joinType.isInstanceOf[ExistenceJoin]) {
+        inputStreamedOutput.indices.map(ExpressionBuilder.makeSelection(_)) ++
+          Seq(ExpressionBuilder.makeSelection(streamedOutput.size))
+      } else {
+        leftOutput.indices.map(ExpressionBuilder.makeSelection(_)) ++
+          rightOutput.indices.map(idx => ExpressionBuilder.makeSelection(idx + streamedOutput.size))
+      }
+    }
+
+    RelBuilder.makeProjectRel(
+      joinRel,
+      new java.util.ArrayList[ExpressionNode](resultProjection.asJava),
+      createExtensionNode(
+        if (exchangeTable) getDirectJoinOutputSeq(joinType, buildOutput, streamedOutput)
+        else getDirectJoinOutputSeq(joinType, streamedOutput, buildOutput),
+        validation),
+      substraitContext,
+      operatorId)
+  }
+
+  def createTransformContext(exchangeTable: Boolean,
+                             output: Seq[Attribute],
+                              rel: RelNode,
+                             inputStreamedOutput: Seq[Attribute],
+                             inputBuildOutput: Seq[Attribute]): TransformContext = {
+    val inputAttributes = if (exchangeTable) {
+      inputBuildOutput ++ inputStreamedOutput
+    } else {
+      inputStreamedOutput ++ inputBuildOutput
+    }
+    TransformContext(inputAttributes, output, rel)
+  }
+}
